html checkbox, databinding on row
Asp.net checkboxes render inside a span, and Knockout's data-bind is
left on the span. There are ways around this, but for our pursposes, an
HTML checkbox with runat="server" is enough.
Besides that, the most important change here is the data-bind="with:
customers()[<%# Container.ItemIndex %>]" on our table row.